examples: Compute goodput in double

+++ examples/sim.cc
@@ -97,10 +97,9 @@ uint64_t LinkConfig::compute_expected_goodput(Timestamp::duration rtt) const {
   constexpr double margin = 0.95;
 
   return std::min(g,
-                  static_cast<uint64_t>(
-                    static_cast<double>(MAX_UDP_PAYLOAD_SIZE * NGTCP2_SECONDS /
-                                        static_cast<uint64_t>(rtt.count())) /
-                    sqrt(loss) * 8 * margin));
+                  static_cast<uint64_t>(MAX_UDP_PAYLOAD_SIZE * NGTCP2_SECONDS /
+                                        static_cast<double>(rtt.count()) /
+                                        sqrt(loss) * 8 * margin));
 }
 
 namespace {
